requirements.txt
Flask
# ----- 必備的 Google Cloud/AI 套件 -----
google-cloud-dialogflow-cx
google-genai # 或 google-cloud-aiplatform, 取決於您使用哪個 SDK 呼叫 Gemini
# ----------------------------------------